+++ game-engine/src/lib.rs
@@ -1,9 +1,6 @@
-#[macro_use]
-extern crate lazy_static;
-
 mod context;
-pub mod physics;
 
+pub use physics;
 pub use rendering;
 
 pub use context::Context;
