+++ physics/Cargo.toml
@@ -1,13 +1,10 @@
 [package]
-name = "game-engine"
+name = "physics"
 version = "0.1.0"
 edition = "2021"
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
-thiserror = "1.0"
-log = "0.4"
-rendering = { path = "../rendering" }
-physics = { path = "../physics" }
 common = { path = "../common" }
+rendering = { path = "../rendering" }
