+++ physics/src/box.rs
@@ -1,10 +1,10 @@
-use crate::physics::{clamp, macros::debug_assert_finite};
+use common::{Mat3, Quaternion, Vec3};
+
 pub(crate) mod collision;
 pub(crate) mod mesh;
 pub(crate) mod sat;
 
-use super::PhysicsMaterial;
-use common::{Mat3, Quaternion, Vec3};
+use crate::{clamp, macros::debug_assert_finite, PhysicsMaterial};
 
 #[derive(Debug, PartialEq, Clone, Copy)]
 pub struct BoxColider {
