+++ res/textures/prototype_512x512_blue2.png
