+++ res/textures/prototype_512x512_cyan.png
