+++ res/textures/prototype_512x512_magenta.png
