+++ res/textures/prototype_512x512_normal.png
