+++ res/textures/prototype_512x512_olive.png
