+++ res/textures/prototype_512x512_pink.png
