+++ res/textures/prototype_512x512_purple.png
