+++ res/textures/prototype_512x512_red.png
