+++ res/textures/prototype_512x512_white.png
