+++ res/textures/prototype_512x512_yellow.png
